# src/classifier/schema/predict_batch.py
# generated by datamodel-codegen:
#   filename:  predict_batch.json
#   timestamp: 2021-08-21T17:26:17+00:00

from __future__ import annotations

from pydantic import BaseModel


class Info(BaseModel):
    total_rows_uploaded: int
    total_rows_w_na: int
    total_rows_added_to_db: int


class Model(BaseModel):
    status: str
    status_code: int
    info: Info